Remove MediaPlayer because it does not make sense to separate
